Fix newline problem on Windows
Parser was missing newlines on Windows files as it was only checking for the line feed character ('\n'). This was causing it to return a PlaintextNode when encountering a carriage return ('\r'), causing all sorts of problems if the white space before the newline character didn't provide the correct level of indentation (such as collapsing all elements on a completely blank line and placing following content after the html element's end tag).